host_agent/pyproject.toml
# This file defines the workspace for the a2a-host-agent project.
# It tells uv that this project is a monorepo containing multiple packages.

[project]
name = "a2a-travel-host-agent" 
version = "0.1.0"
description = "A multi-agent travel planning system with orchestrator and specialized agents."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Shared ADK & A2A Dependencies
    "a2a-sdk==0.2.5",
    "google-adk==1.3.0",
    "nest-asyncio",
    "python-dotenv",
   
    "uvicorn",
    "httpx",

]
